fix: apply code review suggestions for pagination
