+++ src/shared/dtos/index.ts
@@ -1,0 +1,1 @@
+export * from '@shared/dtos/pagination.dto';
