+++ src/shared/interceptors/pagination.interceptor.ts
@@ -7,28 +7,6 @@ import {
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { Response } from 'express';
-import { IsIn, IsInt, IsOptional, Min } from 'class-validator';
-import { Type } from 'class-transformer';
-
-export type OrderBy = 'asc' | 'desc';
-
-export class Pagination {
-  @IsInt()
-  @Min(1)
-  @IsOptional()
-  @Type(() => Number)
-  page: number = 1;
-
-  @IsInt()
-  @IsOptional()
-  @Min(1)
-  @Type(() => Number)
-  limit: number = 10;
-
-  @IsOptional()
-  @IsIn(['asc', 'desc'])
-  orderBy: OrderBy = 'asc';
-}
 
 interface PaginatedResponse<T> {
   data: T;
